Changes suggested in the review code have been added.
