--- adivinaLaPalabra/src/main/java/com/soprasteria/adivinaLaPalabra/AdivinaLaPalabraApplication.java
@@ -1,12 +1,0 @@
-package com.soprasteria.adivinaLaPalabra;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AdivinaLaPalabraApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(AdivinaLaPalabraApplication.class, args);
-	}
-}
