--- adivinaLaPalabra/src/main/java/com/soprasteria/adivinaLaPalabra/dto/PositionOfWordResponse.java
@@ -1,15 +1,0 @@
-package com.soprasteria.adivinaLaPalabra.dto;
-
-import lombok.Data;
-
-@Data
-public class PositionOfWordResponse {
-
-    private char letter;
-    private String hitStatus;
-
-    public PositionOfWordResponse(char letter, String hitStatus) {
-        this.letter = letter;
-        this.hitStatus = hitStatus;
-    }
-}
