--- adivinaLaPalabra/src/main/java/com/soprasteria/adivinaLaPalabra/dto/RoundResponse.java
@@ -1,11 +1,0 @@
-package com.soprasteria.adivinaLaPalabra.dto;
-
-import lombok.Data;
-
-import java.util.UUID;
-
-@Data
-public class RoundResponse {
-
-    private Long id;
-}
