--- adivinaLaPalabra/src/main/java/com/soprasteria/adivinaLaPalabra/model/RoundEntity.java
@@ -1,16 +1,0 @@
-package com.soprasteria.adivinaLaPalabra.model;
-
-import jakarta.persistence.*;
-import lombok.Data;
-
-@Data
-@Entity
-@Table(name = "round")
-public class RoundEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private String word;
-
-}
