--- adivinaLaPalabra/src/main/java/com/soprasteria/adivinaLaPalabra/repository/AllowWords.java
@@ -1,46 +1,0 @@
-package com.soprasteria.adivinaLaPalabra.repository;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.springframework.stereotype.Component;
-import java.io.IOException;
-import java.io.InputStream;
-import java.util.List;
-import java.util.Random;
-
-@Component
-public class AllowWords {
-
-    private List<String> dictionary;
-
-    public AllowWords() {
-        convertJsonToList();
-    }
-
-    private void convertJsonToList() {
-        ObjectMapper objectMapper = new ObjectMapper();
-        try (InputStream inputStream = getClass().getResourceAsStream("/dictionary.json")) {
-            dictionary = objectMapper.readValue(
-                    inputStream,
-                    objectMapper.getTypeFactory().constructCollectionType(List.class, String.class)
-            );
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    public boolean checkWord(String word) {
-        return dictionary.contains(word);
-    }
-
-    public List<String> getDictionary() {
-        return dictionary;
-    }
-
-    public String getRandomWord() {
-        Random random = new Random();
-        int dictionarySize = dictionary.size();
-        int randomIndex = random.nextInt(dictionarySize);
-
-        return dictionary.get(randomIndex);
-    }
-}
