--- adivinaLaPalabra/src/main/java/com/soprasteria/adivinaLaPalabra/repository/RoundRepository.java
@@ -1,9 +1,0 @@
-package com.soprasteria.adivinaLaPalabra.repository;
-
-import com.soprasteria.adivinaLaPalabra.model.RoundEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface RoundRepository extends JpaRepository<RoundEntity, Long> {
-}
