--- adivinaLaPalabra/src/main/java/com/soprasteria/adivinaLaPalabra/service/CheckWordService.java
@@ -1,8 +1,0 @@
-package com.soprasteria.adivinaLaPalabra.service;
-
-
-import com.soprasteria.adivinaLaPalabra.dto.WordResponse;
-
-public interface CheckWordService {
-    WordResponse checkWord(String word, long id);
-}
