--- adivinaLaPalabra/src/main/java/com/soprasteria/adivinaLaPalabra/service/RoundService.java
@@ -1,7 +1,0 @@
-package com.soprasteria.adivinaLaPalabra.service;
-
-import com.soprasteria.adivinaLaPalabra.dto.RoundResponse;
-
-public interface RoundService {
-    RoundResponse newRound();
-}
