+++ adivinaLaPalabra/src/main/java/com/soprasteria/adivinalapalabra/AdivinaLaPalabraApplication.java
@@ -1,0 +1,12 @@
+package com.soprasteria.adivinalapalabra;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class AdivinaLaPalabraApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(AdivinaLaPalabraApplication.class, args);
+	}
+}
