+++ adivinaLaPalabra/src/main/java/com/soprasteria/adivinalapalabra/dto/RoundResponse.java
@@ -1,0 +1,9 @@
+package com.soprasteria.adivinalapalabra.dto;
+
+import lombok.Data;
+
+@Data
+public class RoundResponse {
+
+    private Long id;
+}
