+++ adivinaLaPalabra/src/main/java/com/soprasteria/adivinalapalabra/model/RoundEntity.java
@@ -1,0 +1,16 @@
+package com.soprasteria.adivinalapalabra.model;
+
+import jakarta.persistence.*;
+import lombok.Data;
+
+@Data
+@Entity
+@Table(name = "round")
+public class RoundEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private String word;
+
+}
