+++ adivinaLaPalabra/src/main/java/com/soprasteria/adivinalapalabra/model/enums/HitsStatus.java
@@ -1,0 +1,7 @@
+package com.soprasteria.adivinalapalabra.model.enums;
+
+public enum HitsStatus {
+    HIT,
+    PARTIAL_HIT,
+    FAIL
+}
