+++ adivinaLaPalabra/src/main/java/com/soprasteria/adivinalapalabra/service/RoundService.java
@@ -1,0 +1,7 @@
+package com.soprasteria.adivinalapalabra.service;
+
+import com.soprasteria.adivinalapalabra.dto.RoundResponse;
+
+public interface RoundService {
+    RoundResponse newRound();
+}
