+++ adivinaLaPalabra/src/main/java/com/soprasteria/adivinalapalabra/service/WordService.java
@@ -1,0 +1,8 @@
+package com.soprasteria.adivinalapalabra.service;
+
+
+import com.soprasteria.adivinalapalabra.dto.WordResponse;
+
+public interface WordService {
+    WordResponse checkWord(String word, long id);
+}
