--- adivinaLaPalabra/src/test/java/com/soprasteria/adivinaLaPalabra/AdivinaLaPalabraApplicationTests.java
@@ -1,13 +1,0 @@
-package com.soprasteria.adivinaLaPalabra;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AdivinaLaPalabraApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
