+++ adivinaLaPalabra/src/test/java/com/soprasteria/adivinalapalabra/AdivinaLaPalabraApplicationTests.java
@@ -1,0 +1,13 @@
+package com.soprasteria.adivinalapalabra;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class AdivinaLaPalabraApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
